Rename colorPicker file to MapHelpers
